Correct the byte order of bitfields, don't reverse big-endian fields.
This fixes two issues:
- Bitfields would be passed the data with all bytes reversed vs what
the reads would require, which caused multi-byte bitfields to read
incorrectly, especially if they are not exactly byte-sized.
- Big-endian bitfields would have their AST node children reversed to
read in that order, but this would cause non-byte-sized bitfields to
read aligned to the right byte instead of the left, and conditional
statements would not work at all.