Track bit offsets in evaluator, make bitwise fields sized accurately
Evaluator now stores the bit offset for bitfields that are being instantiated instead of using a callback to track the offset in the `createPatterns()` stack.

In the process making of this change, the bitfields' fields will now use their own position and size instead of the parent bitfield's, since they are not aware of their parent until after all fields are laid out. This results in bitfields displaying the exact bytes they must read from to get their bitwise values, allowing users to see more accurately the location of each field.