parser/evaluator: Enable nesting, arrays, function calls, match and templates in bitfields (#34)
* Fix little-endian bitfield read order

Little-endian bitfields must be read from the least significant bit of the least significant byte. A read would start from the bit:

```
0b00000000'00000000
    here ^
```

continuing to:

```
0b00000000'00000000
  ^ here
```

then to the next byte:

```
0b00000000'00000000
             here ^
```

* Track bit offsets in evaluator, make bitwise fields sized accurately

Evaluator now stores the bit offset for bitfields that are being instantiated instead of using a callback to track the offset in the `createPatterns()` stack.

In the process making of this change, the bitfields' fields will now use their own position and size instead of the parent bitfield's, since they are not aware of their parent until after all fields are laid out. This results in bitfields displaying the exact bytes they must read from to get their bitwise values, allowing users to see more accurately the location of each field.

* Allow bitfields to be nested within other bitfields

An abstract class for patterns that can be bitfield members has been added to allow bitfields to be included as members of other bitfields, and also allows sorting to work within bitwise contexts.

* Implement bitfield member arrays

Currently this only allows other bitfields to be used as the type of the array, but once specific built-in types or enums are allowed for bitfield members, those can be used in arrays as well.

* Add `current_bit_offset` and `read_bits` to `builtin::std::mem`

* Parse function calls within bitfields

* Enable local variables within bitfields

* Parse control flow statements within bitfields

* Allow parsing of `match` statements in bitfields

This unifies the parsing of both `if` and `match` statements between structs and bitfields, so that each one can choose what function is called to parse its members within the conditional blocks.

* Parse template parameters for bitfields

* Small style adjustments

---------

Co-authored-by: Nik <werwolv98@gmail.com>