Properly reference count pattern-local variables
Local variables were being ref counted starting at 0, meaning that if by any process a pattern was created accessing one, it would increment the refcount and then decrement it, causing the variable to be deleted incorrectly.

Now, setting the offset and the section ID will both "destroy" their pattern and then "create" a new one according to the evaluator, allowing it to track whether all patterns are currently referring to a local.