Allow setting bitfield order with `[[bitfield_order(direction, size)]]`
This removes the existing bitfield attributes, pragmas and settings in favor of a new attribute that specifies the direction of a bitfield's layout order as starting from either the least significant or most significant bits. In order to facilitate this, the attribute requires a size parameter in bits to be passed, so that if the byte-endianness defines that the starting point is at the last byte in memory, the bitfield can be offset to start there.