feat: added undef preprocessor directive. fix: missing char in message of lexer. fix: preprocessor no longer does replacements inside strings or doc comments. Also defines are only applied after their position in the code instead of globally. Redefinition of macros is allowed but a warning message is issued. Defines are only allowed to use non-empty and valid identifiers as macro names. fix: problem that directives without complete syntax did not generate error messages. (as in #define with nothing after last e) fix: some combinations  of comments were not handled correctly for example /**/ was parsed as a beginning doc comment. fix: nested #ifdef directives were not parsed correctly producing errors elsewhere. fix: The deletion of comments by preprocessor resulted in wrong column numbers being displayed in error messages. fix: two consecutive directives with no space between them caused the failure to parse the second directive.
Needs work: if an error occurs inside an included file, its location is reported as it is in the preprocessed code. Since the entire file is stored in a single line to preserve line numbers in including pattern this results in making it hard to track the error. The reported column number needs to be converted to the correct position of the error in the file. Better yet would be to add references to the file being included so they can be parsed in their original form and generate better errors.