preprocessor/lexer: Added #undef and handle preprocessing on a token level (#77)
* feat: added undef preprocessor directive.
fix: missing char in message of lexer.
fix: preprocessor no longer does replacements inside strings or doc comments. Also defines are only applied after their position in the code instead of globally. Redefinition of macros is allowed but a warning message is issued. Defines are only allowed to use non-empty and valid identifiers as macro names.
fix: problem that directives without complete syntax did not generate error messages. (as in #define with nothing after last e)
fix: some combinations  of comments were not handled correctly for example /**/ was parsed as a beginning doc comment.
fix: nested #ifdef directives were not parsed correctly producing errors elsewhere.
fix: The deletion of comments by preprocessor resulted in wrong column numbers being displayed in error messages.
fix: two consecutive directives with no space between them caused the failure to parse the second directive.

Needs work: if an error occurs inside an included file, its location is reported as it is in the preprocessed code. Since the entire file is stored in a single line to preserve line numbers in including pattern this results in making it hard to track the error. The reported column number needs to be converted to the correct position of the error in the file. Better yet would be to add references to the file being included so they can be parsed in their original form and generate better errors.

* feat: Added error message for macro redefinition. We actually allow macro redefinition,but the implementation requires a warning to be issued with the output when a macro already defined is given a new definition and since there is no support for issuing warnings a decision was made to set it to an error for now and change it when warnings become available.

* improv: Moved lexing to occur before preprocessing which provides the following advantages:
1) Only one lexical analysis tool to maintain.
2) Tokens that refer to un-preprocessed source code with correct line and column values. This allows ImHex to display error messages using the original source code.
3) Preprocessing tokens makes directives easier to implement, maintain and expand. For example adding function macros with variadic arguments.
4) Errors that occur inside included files can now be traced.
5) Semantic analysis done by the parser now becomes accessible to the syntax highlighter because both use the same tokens.
6) Simplified pl evaluation workflow.
In contrast the disadvantages are few namely an slight increase in lexical complexity and the imposition of language tokenization.

* fix: include directives were being detected incorrectly

* fixed compilation errors

* Just like with the changes to fix imports with templates, merged WerWolv's changes to pattern_language incorrectly on github so builds failed. Brought changes to local clone and fixed compilation errors again.

* Changes according to pr review. Renamed all unsigned to u32, use token comparison instead of strings, removed unnecessary name resolutions, removed unnecessary functions.