fix: the debugger uses locations to set the line of the text editor where the next break will occur. The value returned by getPauseLine() is the number of the line above the one it wants, but the values it may return range from zero to lines-1. When the code that's being executed cannot be seen because it is in another file, the debugger sets the line to zero which in turn is adjusted to line number by subtracting one. As negative line number locations are asserted against getPauseLine() now makes sure that the smallest number it can return is one.
fix: found a bug where multi variable definitions were being label incorrectly by the parser which caused the highlighter to ignore all instances of the variables. Due to the name of the function involved it  appeared that only custom type member definitions were being processed, but it appears the function variable definitions were processed by the code. In order to tell one set of variables from the other an specific node variable that only exists in one of the two provides the litmus test.