fix: Using the windows path delimiter \ results in error.
This can be an issue if you are copying the path directly from explorer. Currently, the old highlighter code changes the color of the char but that will change with the new code. The fix was to add a parameter to the function that processes the directives in the lexer to allow it to skip the escape check.