fix: If last line is a directive, EoP may be  removed from token sequence.
This happens because the new nextline() function skips tokens. When the directive is a define, the end of program was being used as the value for the macro defined.

This fix makes sure that EoP is checked for in nextLine and if found then it is added to the token sequence. Also, if the EoP is added as value for define, then it is removed the token sequence is rewound.