fix: remaining issues with breakpoints. Forgot to reset the line lengths and the lastPausedLine variables on every new evaluation.lastPauseLine is necessary if e.g. You press f5 while paused at a breakpoint so the evaluation hits the breakpoint again.
More importantly the new code now handles one-step debug a lot better than before. The way nodes get coordinates associated with them makes it impossible touse line number information to determine when the next pause will occur. That was fixed for breakpoints, and the new code also fixes it for one-step debugging as well. Additionally, only source code will be subject to pause. Previously, if you placed a breakpoint at a location that happened to be evaluated in an imported file, it was hitting the imported file too. Also stepping through code would go inside imported files which is not what we want most likely given the fact that only source code is displayed while debugging.