preprocessor: Directives arguments are the only tokens allowed in the same line as the directive except comments and end of program. (#125)
* fix: Directives arguments are the only tokens allowed in the same line as the directive except comments and end of program.

Added a function that will ignore any tokens remaining after the directive has been fully parsed.

* fix: If last line is a directive, EoP may be  removed from token sequence.

This happens because the new nextline() function skips tokens. When the directive is a define, the end of program was being used as the value for the macro defined.

This fix makes sure that EoP is checked for in nextLine and if found then it is added to the token sequence. Also, if the EoP is added as value for define, then it is removed the token sequence is rewound.