preprocessor/parser: Unified pragma once system for imports and includes. (#133)
* improv: Unified pragma once system for imports and includes.

fix: certain combinations of mixing imports and includes could cause strange and unexpected errors. A large part of the problem was that a file that was both included and imported ended up as two files with different ids.

improv: includes now support the dot style path specification that import uses.

fix: fixed bug that pr 128 of the pattern language repo claimed to have fixed but actually did not. The bug was that if a file that defined no types but instantiated some variable was included or imported twice, then a duplicate variable error will be generated. The error was created when, after checking if file was included once, The AST was generated because the size of the types was zero. The AST contains the instantiations of variables which are duplicated generating the error.

In terms of data types there were two major changes. The first is how source ids are calculated. The old method of increasing a static variable doesn't guarantee that the same path always leads to the same id. The file ids are now calculated using crc32 on the full path file name which turns the uncertainty into certainty.

The second change involves the variable types for onceIncluded and onceIncludedFiles. Instead of storing once included files in sets of integers, include files now use the same data type than imports uses.

A file that is imported earlier that has pragma once cab block the inclusion of the same file later on. so we need to record which files are imported during preprocessing but only to determine what includes, if any, are affected by them.
To allow includes to block imports, we use the once included files list generated during preprocessing to place entries in the once included file list generated during parsing when the tokens that resulted from the inclusion are being parsed.This way subsequent imports of the same file will be blocked by the includes.
To fix the bug mentioned in pr #128 we remove the check of an empty types vector and returns what can possibly be an empty string.
These changes were tested using a file that imported/included another file which was imported/included first. The tests checked all combinations of import vs. include both with and without  pragma once and produced expected results in all cases.
Some changes in this pr are also part of prs that have been waiting to be reviewed for a while now in order to avoid conflicts that would occur regardless of the order in which they are merged.

* fix: Stupid mistake made all ids equal with value of zero, so none of the tests worked all of a sudden

* Various fixes

---------

Co-authored-by: Nik <werwolv98@gmail.com>