libstd: std::Accumulate math function didn't work. (#135)
The function used to perform operations on sets of data only worked for 8 bit data types. The number of operations was set correctly, but the pointer only incremented one byte at a time.

The pointer now advances the size of the data type in bytes and the end condition of the loop is changed accordingly.