parser: Changes to make syntax highlighting a bit easier. (#127)
Namespaces that are found in included files are not labeled by parser.  Look for the namespace keyword to find all. Locations excluded by ifdef cannot repeat, and it doesn't make sense to validate them, so I removed the functions that were appending and validating them.
Initially syntax highlighting needed to have access to two set of tokens. Un-preprocessed tokens contain directives and comments that need to be highlighted. Preprocessed tokens are used by the parser which adds identifier types to variable definitions.
Instead of passing two very similar sets of tokens, the original ones are merged into one by assigning the information added by the parser to the un-preprocessed tokens.
Other changes include corrections to arguments missing reference operators and name changes to un-shadow names of inherited functions. It should be self-explanatory.