Error markers we displayed in the wrong place.
The problem existed for directives and the import statement. For example omitting the file to import or include resulted in an error marker located in the next line. The problem was that the tokens were being advanced prematurely due to a previous error fix that occurred when the same file was included twice. Imports suffered the same problem cause they were added after the bug fix. A pattern that included the same file twice was run, and it produced no errors.
Other fixes and improvements include:
1) Moved the calculation of the longest line length of pattern source code into the lexer which already reads all the lines and can obtain it with little extra code. The code to use this improvement will be added in a soon to come pr for ImHex.
2) The preprocessor still had code to remove tabs and extra newline characters that was both incorrect and is now unnecessary since the text editor is handling that directly. To test this code import the PE pattern which uses tabs instead of spaces