fix: Incorrect horizontal scrollbar displayed. (#160)
* fix: Incorrect horizontal scrollbar displayed.

After the recent changes to scrolling the code uses the longest line in the source code to determine if and how big horizontal scrollbars are needed. Instead of the computationally expensive calculation of the longest line, we let the lexer record it when it generates the token stream. It turns out that the lexer is also getting token streams for the included and imported files which leads to the problem.

Instead of letting the lexer do both the recording and the reporting of the longest line in the code we use the preprocessor to intercept the value before the included and imported files are preprocessed.

The fix is incomplete unless an upcoming PR to ViewPatterneditor.cpp where the longest line size is sent to the text editor is also  included.

* Update fmt submodule to the latest version