fix: values of local variables that are children of structs, arrays, bitfields or unions are set to invalid values when the setOffset() function of the parent is called. (#174)
For example when a child of the Json built in data type is used as a reference argument to a function the values of the child and all its children become zero. For those four pattern types `setOffset()` changes the offsets of all the children and if the children are local the offset is set to the offset of the parent pattern which is clearly wrong. The fix consists on removing the code that changes the offsets of local variables in the four (five?) setOffset() functions.

 Also fixed unions using local variables to determine the pattern size.